Hoisted Main Screen, Extracted some objects into their own composables.
