+++ app/src/main/java/com/example/studyapp/ui/viewmodel/CurrentQuestionViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.studyapp.ui.viewmodel
+
+class CurrentQuestionViewModel {
+}
